rename `graph` to `control_flow_graph::implementation`
